refactor if statement 중괄호 추가
